+++ src/components/layout/MainLayout.tsx
@@ -1,7 +1,8 @@
 'use client';
-import GlobalPlusButton from '@/components/layout/GlobalPlusButton';
+
 import Header from '@/components/layout/Header';
 import Sidebar from '@/components/layout/sidebar/Sidebar';
+import MemoCreateButton from '@/components/memo-editor/MemoCreateButton';
 import { useState } from 'react';
 
 interface MainLayoutProps {
@@ -14,6 +15,7 @@ export default function MainLayout({ children }: MainLayoutProps) {
   const toggleSidebar = () => {
     setIsSidebarOpen(prev => !prev);
   };
+
   return (
     <div className="flex flex-row min-h-screen relative">
       <div className="flex-1 flex flex-col min-h-screen w-full md:w-auto">
@@ -21,7 +23,7 @@ export default function MainLayout({ children }: MainLayoutProps) {
         <main className="flex-1 container mx-auto px-4 py-3 mt-16">{children}</main>
       </div>
       <Sidebar isOpen={isSidebarOpen} onToggle={toggleSidebar} className="md:static absolute" />
-      <GlobalPlusButton />
+      <MemoCreateButton />
     </div>
   );
 }
